requirements.txt
pyaudio
pvporcupine
whisper
llama-cpp-python
torch
soundfile
piper-tts
